pyproject.toml
[project]
name = "skel-helper-bot"
version = "0.1.0"
requires-python = ">=3.12"
description = "Skel Helper Bot connects Telegram chats to the Skel Crypto Agent over HTTP."
dependencies = [
  "python-telegram-bot>=20.8,<21.0",
  "httpx>=0.26,<0.27",
  "python-dotenv>=1.1,<2.0",
  "loguru>=0.7,<0.8",
  "python-ulid>=3.0,<4.0",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
